22-07-28 algorithm practice
