+++ leetCode/Easy/1046. Last_stone_weight.py
