+++ leetCode/Easy/1342. Number_of_steps_to_reduce_a_number_to_zero.py
