+++ leetCode/Easy/1365. How_many_numbers_are_smaller_than_the_current_number.py
