+++ leetCode/Easy/1450. Number_of_students_doing_homework_at_a_given_time.py
