+++ leetCode/Easy/1528. Shuffle_string.py
