+++ leetCode/Easy/1572. Matrix_diagonal_sum.py
