+++ leetCode/Easy/1598. Crawler_log_folder.py
