+++ leetCode/Easy/228. Summary_ranges.py
