+++ leetCode/Easy/700. Search_in_a_binary_search_tree.py
