+++ leetCode/Easy/821. Shortest_distance_to_a_character.py
