+++ leetCode/Easy/867. Transpose_matrix.py
