+++ leetCode/Easy/897. Increasing_order_search_tree.py
