+++ leetCode/Medium/1641. Count_sorted_vowel_strings.py
