+++ leetCode/Medium/701. Insert_into_a_binary_search_tree.py
